requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
click==8.2.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.116.2
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.11.3
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
sentry-sdk==2.38.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
